Modified openmpt123 sources with new --pattern-time-log argument to support the Houdini importer.
I wasn't careful enough about preserving existing behavior so to have any chance of contributing this back to the OpenMPT project this would need to be carefully re-merged with the originals so that nothing changes about the existing pattern-display mode. Totally doable, but I have no idea whether it would be accepted even if I did this so I'm holding off for now. (but still sharing the sources here in case anyone else wants to play!)

I have tried to communicate with the OpenMPT devs about this but they haven't responded much so far: https://forum.openmpt.org/index.php?topic=7053